// web/src/api/generated/model/githubHealthResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Condux Control Plane API
 * OpenAPI spec version: v1
 */

export interface GithubHealthResponse {
  status: string;
  /** @nullable */
  installationId: number | null;
  /** @nullable */
  accountLogin: string | null;
  /** @nullable */
  detail: string | null;
}